Add style for new rubber duck debugging link
Set the text-decoration to none and the color to black for when the 'Rubber Duck Debugging' link is not in use, so that the title appears normal. For the hover and active states of the link the color is set to blue and the text-decoration is underline so that it's clear it is actually a link.